refactor: TopをTopとTopContainerに分割
