+++ frontend/src/Top.css
@@ -9,10 +9,3 @@
     align-items: center;
     justify-content: center;
 }
-
-.card-image {
-    width: 30vh;
-    height: auto;
-    border-radius: 4px;
-    margin: 40px;
-}
