+++ frontend/src/TopContent.css
@@ -1,0 +1,6 @@
+.card-image {
+    width: 30vh;
+    height: auto;
+    border-radius: 4px;
+    margin: 40px;
+}
